requirements.txt
anthropic>=0.40.0
click>=8.0
rich>=13.0
pyyaml>=6.0
python-dotenv>=1.0
